pyproject.toml
[project]
name = "mock-paper-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crewai>=0.203.1",
    "fastmcp>=2.12.4",
    "langchain-chroma>=0.2.6",
    "langchain-community>=0.3.31",
    "langchain-google-genai>=2.1.12",
    "markdown-pdf>=1.10",
    "pypdf>=6.1.1",
    "python-dotenv>=1.1.1",
    "streamlit>=1.50.0",
]
